Move independent concepts out of scheduler plugin in their own files
